refactor(shell): reorganize header components
Moved NavLink and ThemeSwitch components out of the header directory
to better align with the app structure. Updated import paths
accordingly for more maintainable and scalable code organization.
