+++ apps/web/app/shell/header.tsx
@@ -8,8 +8,8 @@ import { clsx } from '@suddenlygiovanni/ui/lib/utils.ts'
 
 import avatarAssetUrl from '../assets/giovanni_ravalico-profile_bw.webp'
 import { routesRecord } from '../routes-record.ts'
-import { NavLink } from './header/nav-link.tsx'
-import { ThemeSwitch } from './header/theme-switch.tsx'
+import { NavLink } from './nav-link.tsx'
+import { ThemeSwitch } from './theme-switch.tsx'
 
 const routes = (
 	[
